+++ calculator/index.js
@@ -1,22 +1,3 @@
-function add(a, b) {
-	return a + b;
-}
-
-function subtract(a, b) {
-	return a - b;
-}
-
-function multiply(a, b) {
-	return a * b;
-}
-
-function divide(a, b) {
-	if (b === 0) {
-		throw new Error('Cannot divide by zero');
-	}
-	return a / b;
-}
-
 let result = 0;
 let operation = '';
 let input1 = document.getElementById('input1');
